+++ skyflash/utils.py
@@ -8,11 +8,45 @@ import enum
 import traceback
 import subprocess
 import json
+import ipaddress
 from PyQt5.QtCore import QObject, pyqtSignal, QRunnable, pyqtSlot
 
 if 'nt' in os.name:
     import wmi
     import win32file
+
+def cleanString(data):
+    '''Cleans a string from trailing or leading chars
+
+    The list of chars is the cleanOf below, it's just a hack with strip()
+    but saves you from typing boring sentences over and over'''
+
+    cleanOf = " ,.-"
+
+    return data.strip(cleanOf)
+
+def splitDNS(dnsString):
+    '''Split a string that can represent up to tree DNS entries
+    return a list with the entries, if on error first entry is False
+    and second is the reason
+    '''
+
+    # first detecting the split pattern and detecting the most likely
+    data = dnsString.split(" ")
+    coma = dnsString.split(",")
+    if len(coma) > len(data):
+        data = coma
+
+    dns = []
+    err = False
+    for e in data:
+        result, ip = (validIP(cleanString(e)))
+        if result:
+            dns.append(str(ip))
+        else:
+            return [False, ip]
+
+    return dns
 
 def shortenPath(fullpath, ccount):
     '''Shorten a passed FS path to a char count size'''
@@ -114,24 +148,18 @@ def size(size):
 
     return out
 
-def validIP(ip):
+def validIP(iip):
     '''Takes a string of a IP and return a tuple:
     true/false and a reason if false
+
+    It uses ipaddress stock module, and the error is returned as the default comment 
     '''
 
-    # four digits
-    digits = str(ip).split('.')
-    if len(digits) != 4:
-        return (False, "Not enough digits on the IP")
-
-    # from 0 to 255
-    for d in digits:
-        d = int(d)
-        if d > 255 or d < 0:
-            return (False, "One of the digits on the IP is not valid")
-
-    # all good
-    return (True, "")
+    try:
+        ip = ipaddress.IPv4Address(iip)
+        return (True, ip)
+    except ValueError as messg:
+        return (False, str(messg))
 
 def getLinuxPath(soft):
     '''Make use of getDataFromCLI
@@ -222,7 +250,8 @@ def windowsDevices():
     data = []
 
     for physical_disk in c.Win32_DiskDrive():
-        if 7 in physical_disk.Capabilities:
+        capas = physical_disk.Capabilities
+        if capas != None and 7 in capas:
             # is a removable media
             phy = physical_disk.DeviceID
             size = int(physical_disk.Size)
@@ -238,6 +267,40 @@ def windowsDevices():
             data.append([phy, desc, size, drives])
 
     return data
+
+def linuxMediaDevices():
+    ''' List the media devices that are removable in liunux
+
+    If the major number is 8, that indicates it to be a disk device.
+
+    The minor number is the partitions on the same device:
+    - 0 means the entire disk
+    - 1 is the primary
+    - 2 is extended
+    - 5 is logical partitions
+    The maximum number of partitions is 15.
+
+    Use `$ sudo fdisk -l` and `$ sudo sfdisk -l /dev/sda` for more information.
+    '''
+
+    with open("/proc/partitions", "r") as f:
+        devices = []
+
+        for line in f.readlines()[2:]:
+            words = [ word.strip() for word in line.split() ]
+            major_number = int(words[0])
+            minor_number = int(words[1])
+            device_name = words[3]
+
+            # disk devices by device name, not partitions
+            if (major_number == 8) and not (minor_number % 16):
+                devices.append("/dev/" + device_name)
+
+        # data return
+        return devices
+
+    # default return if devices are not found
+    return False
 
 def getMacDriveInfo():
     '''Get drives info in MacOS
@@ -345,16 +408,15 @@ def getLinDrivesInfo():
     '''Return a list of available drives in linux
     if possible with a drive label and sizes on bytes:
 
-    On some linux the SD cards take /dev/sdb and on
-
     [
-        ('/dev/mmcblk0', '', 2369536),
-        ('/dev/mmcblk1', 'uSD cars 16GB', 3995639808),
-        ('/dev/sda', '', 8300555)
+        ('/dev/mmcblk0', '', 8388608),
+        ('/dev/mmcblk1', 'BIG uSD Card', 16777216),
+        ('/dev/sda', 'MULTIBOOT', 8300555)
     ]
     '''
 
-    drives = listMediaDevices()
+    # get all removable media devices in the system
+    drives = linuxMediaDevices()
 
     # if we detected a drive, gather the details via lsblk
     if drives:
@@ -377,16 +439,17 @@ def getLinDrivesInfo():
     # getting data, output format is: [(drive, "LABEL", total),]
     for device in data['blockdevices']:
         if device['rm'] == '1':
+            name = device['name']
             cap = int(device['size'])
             mounted = ""
             for c in device['children']:
-                if c['mountpoint']:
-                    mounted += " ".join([c['mountpoint'].split("/")[-1]])
+                nn = 'No Label'
+                if c['mountpoint'] is not None:
+                    nn = c['mountpoint'].split("/")[-1]
 
-            if len(mounted) <= 0:
-                mounted += "Not in use"
+                mounted += "{}, ".format(nn)
 
-            finalDrives.append((device['name'], mounted, cap))
+            finalDrives.append((name, mounted.strip(" ,"), cap))
 
     # final test
     if len(finalDrives) > 0:
@@ -395,7 +458,6 @@ def getLinDrivesInfo():
         return False
 
 # fileio overide class to get progress on tarfile extraction
-# TODO How to overide a class from a module
 class ProgressFileObject(io.FileIO):
     '''Overide the fileio object to have a callback on progress'''
 
@@ -432,38 +494,6 @@ class WorkerSignals(QObject):
 
     def __init__(self, parent=None):
         super(WorkerSignals, self).__init__(parent=parent)
-
-def listMediaDevices():
-    ''' List the media devices that are removable
-
-    If the major number is 8, that indicates it to be a disk device.
-
-    The minor number is the partitions on the same device:
-    - 0 means the entire disk
-    - 1 is the primary
-    - 2 is extended
-    - 5 is logical partitions
-    The maximum number of partitions is 15.
-
-    Use `$ sudo fdisk -l` and `$ sudo sfdisk -l /dev/sda` for more information.
-    '''
-
-    with open("/proc/partitions", "r") as f:
-        devices = []
-
-        for line in f.readlines()[2:]:
-            words = [ word.strip() for word in line.split() ]
-            minor_number = int(words[1])
-            device_name = words[3]
-
-            if (minor_number % 16) == 0:
-                path = "/sys/class/block/" + device_name
-
-                if os.path.islink(path):
-                    if os.path.realpath(path).find("/usb") > 0:
-                        devices.append("/dev/" + device_name)
-
-        return devices
 
 # Generic worker to use in threads
 class Worker(QRunnable):